rename `transfer.rs` to `transfer_without_fee.rs` to satisfy clippy
