+++ zk-token-sdk/src/instruction/transfer/mod.rs
@@ -1,10 +1,12 @@
-pub mod transfer;
 pub mod transfer_with_fee;
+pub mod transfer_without_fee;
 
 pub use {
-    transfer::{TransferAmountEncryption, TransferData, TransferProofContext, TransferPubkeys},
     transfer_with_fee::{
         FeeEncryption, FeeParameters, TransferWithFeeData, TransferWithFeeProofContext,
         TransferWithFeePubkeys,
+    },
+    transfer_without_fee::{
+        TransferAmountEncryption, TransferData, TransferProofContext, TransferPubkeys,
     },
 };
