+++ zk-token-sdk/src/instruction/transfer/transfer_with_fee.rs
@@ -10,7 +10,8 @@ use {
         errors::ProofError,
         instruction::{
             combine_lo_hi_ciphertexts, combine_lo_hi_commitments, combine_lo_hi_openings,
-            combine_lo_hi_u64, split_u64, transfer::transfer::TransferAmountEncryption, Role,
+            combine_lo_hi_u64, split_u64, transfer::transfer_without_fee::TransferAmountEncryption,
+            Role,
         },
         range_proof::RangeProof,
         sigma_proofs::{
